LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-anchor-positioning-004-expected.txt


FAIL CSS Content Visibility: auto not skipped, but skipped after removing all anchors assert_equals: #positioned should be anchored because #anchor is on-screen. expected 209 but got 8
